requirements.txt
streamlit
langchain
langchain-google-genai
chromadb
pdfplumber
python-dotenv
arxiv
requests